+++ bijection-core/src/main/scala/com/twitter/bijection/StringBijections.scala
@@ -20,6 +20,7 @@ import java.net.URL
 import java.util.UUID
 
 import scala.annotation.tailrec
+import scala.collection.generic.CanBuildFrom
 
 import Bijection.build
 
@@ -47,6 +48,8 @@ object StringCodec extends StringBijections
  * in Config maps.
  */
 object StringJoinBijection {
+  val DEFAULT_SEP = ":"
+
   @tailrec
   private def split(str: String, sep: String, acc: List[String] = Nil): List[String] = {
     str.indexOf(sep) match {
@@ -56,18 +59,51 @@ object StringJoinBijection {
     }
   }
 
-  def apply(separator: String = ":") =
-    Bijection.build[Iterable[String], Option[String]] { xs =>
-    // TODO: Instead of throwing, escape the separator in the encoded string.
-      assert(!xs.exists(_.contains(separator)), "Can't encode strings that include the separator.")
-      if (xs.isEmpty)
-        None
-      else
-        Some(xs.mkString(separator))
-    } { strOpt => strOpt match {
+  def apply(separator: String = DEFAULT_SEP): Bijection[Iterable[String], Option[String]] =
+    new Bijection[Iterable[String], Option[String]] {
+      override def apply(xs: Iterable[String]) = {
+        // TODO: Instead of throwing, escape the separator in the encoded string.
+        assert(!xs.exists(_.contains(separator)), "Can't encode strings that include the separator.")
+        if (xs.isEmpty)
+          None
+        else
+          Some(xs.mkString(separator))
+      }
+      override def invert(strOpt: Option[String]) =
+        strOpt match {
           case None => Iterable.empty[String]
           // String#split is not reversible, and uses regexs
           case Some(str) => StringJoinBijection.split(str, separator)
         }
     }
+
+  /** Convert a collection of numbers to and from a string
+   * It's common to have types which we know have at least 1 character in their string
+   * representation. Knowing that the empty string is not allowed we can map that to the empty
+   * collection:
+   */
+  def nonEmptyValues[N, B <: TraversableOnce[N]](separator: String = DEFAULT_SEP)
+  (implicit bij: Bijection[N, String], ab: CanBuildFrom[Nothing, N, B]): Bijection[B, String] =
+    Bijection.toContainer[N, String, B, Iterable[String]]
+      .andThen(apply(separator))
+      .andThen(Bijection.filterDefault("").inverse)
+
+  /**
+   * Converts between any collection of A and and Option[String],
+   * given an implicit Bijection[A,String]. To get the final string out,
+   * compose with the getOrElse bijection:
+   *
+   * viaContainer[Int,Set[Int]] andThen Bijection.getOrElse(""): Bijection[Set[Int],String]
+   *
+   * Note that this can be dangerous with empty collections,
+   * as Bijection will try to convert "" -> Int. It's safer to use
+   * an instance of type A with the "as" notation for a default item
+   * in the collection:
+   *
+   * import Bijection.asMethod
+   * viaContainer[Int,Set[Int]] andThen Bijection.getOrElse(0.as[String]): Bijection[Set[Int],String]
+   */
+  def viaContainer[A, B <: TraversableOnce[A]](separator: String = DEFAULT_SEP)
+  (implicit bij: Bijection[A, String], ab: CanBuildFrom[Nothing, A, B]): Bijection[B, Option[String]] =
+    Bijection.toContainer[A, String, B, Iterable[String]] andThen apply(separator)
 }